requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.0
alembic>=1.13.0
# Optional production deps:
# - Postgres driver
psycopg[binary]>=3.1.19
numpy>=1.26.0
orjson>=3.10.0
python-multipart>=0.0.9
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0
Pillow>=10.0.0
qrcode>=7.4.2
imageio>=2.34.0
imageio-ffmpeg>=0.5.0
# Optional S3/MinIO storage backend
boto3>=1.34.0

ray[rllib]>=2.9.0
pettingzoo>=1.24.0
gymnasium>=0.29.0
# CPU-only PyTorch to avoid CUDA init crashes in WSL/sandbox.
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.9.1+cpu

pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.6.0
